+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/pojo/AllocationProductBatch.java
@@ -5,6 +5,8 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @ApiModel("调拨商品批次表")
 @Data
 public class AllocationProductBatch extends CommonBean {
@@ -51,16 +53,16 @@ public class AllocationProductBatch extends CommonBean {
     private Long inventory;
 
     @ApiModelProperty("税率")
-    private Long tax;
+    private BigDecimal tax;
 
     @ApiModelProperty("含税成本")
-    private Long taxPrice;
+    private BigDecimal taxPrice;
 
     @ApiModelProperty("数量")
     private Long quantity;
 
     @ApiModelProperty("含税总成本")
-    private Long taxAmount;
+    private Double taxAmount;
 
     @ApiModelProperty("图片地址")
     private String pictureUrl;
