+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/pojo/MovementProduct.java
@@ -5,6 +5,7 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @ApiModel("移库商品表")
@@ -53,16 +54,16 @@ public class MovementProduct extends CommonBean {
     private Long inventory;
 
     @ApiModelProperty("税率")
-    private Long tax;
+    private BigDecimal tax;
 
     @ApiModelProperty("含税成本")
-    private Long taxPrice;
+    private BigDecimal taxPrice;
 
     @ApiModelProperty("数量")
     private Long quantity;
 
     @ApiModelProperty("含税总成本")
-    private Long taxAmount;
+    private BigDecimal taxAmount;
 
     @ApiModelProperty("删除标记，0未删除 1已删除")
     private Byte delFlag;
