+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/response/movement/QueryMovementResVo.java
@@ -5,6 +5,7 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -61,7 +62,7 @@ public class QueryMovementResVo {
     private Long  callInQuantity;
 
     @ApiModelProperty("含税库存成本，展示时需要除以100")
-    private Long taxInventoryCost;
+    private BigDecimal taxInventoryCost;
 
     @ApiModelProperty("出库单号")
     private String outboundOderCode;
