+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/response/scrap/ScrapResVo.java
@@ -8,6 +8,7 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -57,10 +58,10 @@ public class ScrapResVo {
     private Long callOutQuantity;
 
     @ApiModelProperty("含税总成本")
-    private Long totalCostRate;
+    private BigDecimal totalCostRate;
 
     @ApiModelProperty("含税总成本")
-    private Long callOutTotalCostRate;
+    private BigDecimal callOutTotalCostRate;
 
     @ApiModelProperty("出库单号")
     private String outboundOderCode;
