+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/response/sku/oms/OmsProductSkuItemResp.java
@@ -5,6 +5,7 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -25,7 +26,7 @@ public class OmsProductSkuItemResp {
 
     @ApiModelProperty("动销价")
     @JsonProperty("sku_price")
-    private Long skuPrice;
+    private BigDecimal skuPrice;
 
     @ApiModelProperty("库存")
     @JsonProperty("sku_stock")
